Add get_backup_type function
This commit adds a new function which is used to identify which type of
backup should be run. The way this script is meant to be used is via
cron and it is difficult to set cron up to handle cases like skipping
'lower' order backups when 'higher' order backups are also scheduled.
I.e., if a daily backup is scheduled, then any hourly or minutely
backups scheduled to run should be skipped. This naive solution requires
cron to write a 'backup_schedule' file that indicates the type of the
currently scheduled job. So, if cron is running a daily backup, it will
likely write out the backup_schedule.{minutely,hourly,daily} files and
this script will grab all these files and only run the highest order
backup. This is not perfect, but does seem to work the way I want it to.