Change the way the previous backup is identified
This commit changes the way the 'previous' backup is selected for use
with the --link-dest rsync option. Now, a symbolic link is created every
time this script gets run that points to the most recent backup.  This
symbolic link is located and used for the --link-dest option instead of
finding the last backup of the same type of current backup as was done
before, i.e., find that latest weekly backup if running a weekly backup
instead of using the actual latest backup which would likely be a
minutely or hourly backup.