Introduce initial version of config file parsing
This commit adds very naive, initial logic used to parse a system
default file which will be used to define stuff like backup target,
source, number of backups to keep, etc. This seems to work based on the
little bit of testing that I have performed thus far but needs a lot of
improvement and honing.